# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyReportCreate": {
        "properties": {
          "blocker": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker"
          },
          "nextAction": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Nextaction",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "reportDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reportdate"
          },
          "status": {
            "$ref": "#/components/schemas/WorkStatus"
          },
          "teamId": {
            "title": "Teamid",
            "type": "string"
          },
          "workItemId": {
            "title": "Workitemid",
            "type": "string"
          },
          "workSummary": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Worksummary",
            "type": "string"
          }
        },
        "required": [
          "teamId",
          "projectId",
          "workItemId",
          "status",
          "workSummary",
          "nextAction"
        ],
        "title": "DailyReportCreate",
        "type": "object"
      },
      "DailyReportRead": {
        "properties": {
          "blocker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "effectiveBlocker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectiveblocker"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastEditedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasteditedat"
          },
          "nextAction": {
            "title": "Nextaction",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "reportDate": {
            "format": "date",
            "title": "Reportdate",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkStatus"
          },
          "submittedAt": {
            "format": "date-time",
            "title": "Submittedat",
            "type": "string"
          },
          "teamId": {
            "title": "Teamid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "workItemId": {
            "title": "Workitemid",
            "type": "string"
          },
          "workSummary": {
            "title": "Worksummary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "teamId",
          "userId",
          "projectId",
          "workItemId",
          "reportDate",
          "status",
          "workSummary",
          "blocker",
          "effectiveBlocker",
          "nextAction",
          "source",
          "submittedAt",
          "lastEditedAt",
          "createdAt",
          "updatedAt"
        ],
        "title": "DailyReportRead",
        "type": "object"
      },
      "DailyReportUpdate": {
        "properties": {
          "blocker": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker"
          },
          "nextAction": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextaction"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "workSummary": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worksummary"
          }
        },
        "title": "DailyReportUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkStatus": {
        "enum": [
          "NOT_STARTED",
          "IN_PROGRESS",
          "BLOCKED",
          "READY_FOR_REVIEW",
          "DONE"
        ],
        "title": "WorkStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Project Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/daily-reports": {
      "post": {
        "operationId": "create_daily_report_api_v1_daily_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Daily Report",
        "tags": [
          "daily-reports"
        ]
      }
    },
    "/api/v1/daily-reports/history": {
      "get": {
        "operationId": "list_daily_history_api_v1_daily_reports_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "teamId",
            "required": true,
            "schema": {
              "title": "Teamid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datefrom"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dateto"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DailyReportRead"
                  },
                  "title": "Response List Daily History Api V1 Daily Reports History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Daily History",
        "tags": [
          "daily-reports"
        ]
      }
    },
    "/api/v1/daily-reports/me": {
      "get": {
        "operationId": "list_my_daily_reports_api_v1_daily_reports_me_get",
        "parameters": [
          {
            "in": "query",
            "name": "teamId",
            "required": true,
            "schema": {
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DailyReportRead"
                  },
                  "title": "Response List My Daily Reports Api V1 Daily Reports Me Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Daily Reports",
        "tags": [
          "daily-reports"
        ]
      }
    },
    "/api/v1/daily-reports/options": {
      "get": {
        "operationId": "get_daily_options_api_v1_daily_reports_options_get",
        "parameters": [
          {
            "in": "query",
            "name": "teamId",
            "required": true,
            "schema": {
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": {
                        "type": "string"
                      },
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "title": "Response Get Daily Options Api V1 Daily Reports Options Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Daily Options",
        "tags": [
          "daily-reports"
        ]
      }
    },
    "/api/v1/daily-reports/{report_id}": {
      "put": {
        "operationId": "update_daily_report_api_v1_daily_reports__report_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "required": true,
            "schema": {
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyReportUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Daily Report",
        "tags": [
          "daily-reports"
        ]
      }
    },
...
  }
}